requirements.txt
snowflake-connector-python>=3.0.0
requests>=2.28.0
